+++ api/src/pdc/services/geo/actions/GetPointByAddressAction.ts
@@ -1,13 +1,13 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
+import { GeoProviderInterfaceResolver } from "@/pdc/providers/geo/index.ts";
 import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
 import {
   handlerConfig,
   ParamsInterface,
   ResultInterface,
-} from "@/shared/geo/getPointByAddress.contract.ts";
-import { alias } from "@/shared/geo/getPointByAddress.schema.ts";
-import { GeoProviderInterfaceResolver } from "@/pdc/providers/geo/index.ts";
+} from "@/pdc/services/geo/contracts/getPointByAddress.contract.ts";
+import { alias } from "@/pdc/services/geo/contracts/getPointByAddress.schema.ts";
 
 @handler({
   ...handlerConfig,
