+++ api/src/pdc/services/geo/actions/GetRouteMetaAction.ts
@@ -1,13 +1,9 @@
-import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { handler } from "@/ilos/common/index.ts";
-import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
-import {
-  handlerConfig,
-  ParamsInterface,
-  ResultInterface,
-} from "@/shared/geo/getRouteMeta.contract.ts";
-import { alias } from "@/shared/geo/getRouteMeta.schema.ts";
+import { Action as AbstractAction } from "@/ilos/core/index.ts";
 import { GeoProviderInterfaceResolver } from "@/pdc/providers/geo/index.ts";
+import { hasPermissionMiddleware } from "@/pdc/providers/middleware/index.ts";
+import { handlerConfig, ParamsInterface, ResultInterface } from "@/pdc/services/geo/contracts/getRouteMeta.contract.ts";
+import { alias } from "@/pdc/services/geo/contracts/getRouteMeta.schema.ts";
 
 @handler({
   ...handlerConfig,
