+++ api/src/pdc/services/geo/contracts/getRouteMeta.schema.ts
