+++ api/src/pdc/services/observatory/contracts/distribution/journeysByDistances.contract.ts
@@ -1,9 +1,4 @@
-import {
-  Direction,
-  INSEECode,
-  PerimeterLabel,
-  PerimeterType,
-} from "@/shared/geo/shared/Perimeter.ts";
+import { Direction, INSEECode, PerimeterLabel, PerimeterType } from "@/pdc/services/geo/contracts/shared/Perimeter.ts";
 
 export interface SingleResultInterface {
   code: INSEECode;
@@ -32,5 +27,4 @@ export const handlerConfig = {
   method: "journeysByDistances",
 };
 
-export const signature =
-  `${handlerConfig.service}:${handlerConfig.method}` as const;
+export const signature = `${handlerConfig.service}:${handlerConfig.method}` as const;
