+++ api/src/pdc/services/observatory/contracts/flux/getBestFlux.contract.ts
@@ -1,8 +1,4 @@
-import {
-  INSEECode,
-  PerimeterLabel,
-  PerimeterType,
-} from "@/shared/geo/shared/Perimeter.ts";
+import { INSEECode, PerimeterLabel, PerimeterType } from "@/pdc/services/geo/contracts/shared/Perimeter.ts";
 
 export interface SingleResultInterface {
   territory_1: INSEECode;
@@ -29,5 +25,4 @@ export const handlerConfig = {
   method: "getBestFlux",
 };
 
-export const signature =
-  `${handlerConfig.service}:${handlerConfig.method}` as const;
+export const signature = `${handlerConfig.service}:${handlerConfig.method}` as const;
