+++ api/src/pdc/services/observatory/contracts/flux/getEvolFlux.contract.ts
@@ -1,9 +1,4 @@
-import {
-  IndicType,
-  INSEECode,
-  PerimeterLabel,
-  PerimeterType,
-} from "@/shared/geo/shared/Perimeter.ts";
+import { IndicType, INSEECode, PerimeterLabel, PerimeterType } from "@/pdc/services/geo/contracts/shared/Perimeter.ts";
 
 export interface SingleResultInterface {
   territory: INSEECode;
@@ -33,5 +28,4 @@ export const handlerConfig = {
   method: "getEvolFlux",
 };
 
-export const signature =
-  `${handlerConfig.service}:${handlerConfig.method}` as const;
+export const signature = `${handlerConfig.service}:${handlerConfig.method}` as const;
