+++ api/src/pdc/services/observatory/contracts/flux/getEvolFlux.schema.ts
@@ -1,4 +1,4 @@
-import { indicTypes, perimeterTypes } from "@/shared/geo/shared/Perimeter.ts";
+import { indicTypes, perimeterTypes } from "@/pdc/services/geo/contracts/shared/Perimeter.ts";
 
 export const alias = "observatory.getEvolFlux";
 export const schema = {
