+++ api/src/pdc/services/observatory/contracts/incentive/getIncentive.schema.ts
@@ -1,7 +1,4 @@
-import {
-  directionTypes,
-  perimeterTypes,
-} from "@/shared/geo/shared/Perimeter.ts";
+import { directionTypes, perimeterTypes } from "@/pdc/services/geo/contracts/shared/Perimeter.ts";
 
 export const alias = "observatory.getIncentive";
 export const schema = {
