+++ api/src/pdc/services/policy/engine/helpers/isInside.unit.spec.ts
@@ -1,6 +1,6 @@
 import { assertEquals, it } from "@/dev_deps.ts";
 
-import type { GeoJSON } from "@/shared/geo/GeoJson.ts";
+import type { GeoJSON } from "@/pdc/services/geo/contracts/GeoJson.ts";
 import { StatelessContext } from "../entities/Context.ts";
 import { generateCarpool } from "../tests/helpers.ts";
 import { isInside } from "./isInside.ts";
